+++ SQLiteStudio3/guiSQLiteStudio/multieditor/multieditordialog.h
@@ -17,11 +17,16 @@ class GUI_API_EXPORT MultiEditorDialog : public QDialog
 
         void setValue(const QVariant& value, const DataType& dataType);
         void setValue(const QVariant& value);
-        QVariant getValue();
+        QVariant getValue() const;
+        void setNewRow(bool value);
+        bool isNewRow() const;
+        void setUntouched(bool value);
+        bool isUntouched() const;
 
         void setDataType(const DataType& dataType);
         void setReadOnly(bool readOnly);
         void enableFk(Db* db, SqlQueryModelColumn* column);
+        void configureResetButton(SqlQueryModelColumn* column);
 
     private:
         MultiEditor* multiEditor = nullptr;